refactor: remove test generation pipeline
