+++ src/core/engine/mod.rs
@@ -6,9 +6,6 @@ pub mod baseline;
 pub mod cli_tool;
 pub mod codebase_scan;
 pub mod command;
-pub mod contract;
-pub mod contract_extract;
-pub mod contract_testgen;
 pub mod edit_op;
 pub mod edit_op_apply;
 pub mod execution_context;
