+++ src/core/mod.rs
@@ -16,7 +16,6 @@ pub mod project;
 pub mod refactor;
 pub mod release;
 pub mod rig;
-pub mod scaffold;
 pub mod server;
 pub mod upgrade;
