--- src/core/scaffold/mod.rs
@@ -1,12 +1,0 @@
-pub mod test;
-pub mod workflow;
-
-pub use test::{
-    find_test_location, generated_test_uses_unresolved_types, load_extension_grammar,
-    render_generated_test_append, render_generated_test_scaffold, scaffold_file, scaffold_untested,
-    ExtractedClass, ExtractedMethod, ScaffoldBatchResult, ScaffoldConfig, ScaffoldResult,
-    TestLocation,
-};
-pub use workflow::{
-    run_scaffold_workflow, ScaffoldFileOutput, ScaffoldOutput, ScaffoldWorkflowOutput,
-};
